Move api/src/schema.ts to api/src/schema/schema.ts (to be split out further as it grows). Tests moved from api/src/__tests__ to be co-located with their coresponding source modules
